services/agent/pyproject.toml
[tool.poetry]
name = "ai-agent-platform"
version = "0.1.0"
description = "Local AI agent platform with FastAPI agent server and Typer-based stack CLI."
authors = ["MagTer Contributors <opensource@magter.ai>"]
readme = "README.md"
package-mode = false
packages = [
    {include = "core", from = "src"},
    {include = "orchestrator", from = "src"},
    {include = "interfaces", from = "src"},
    {include = "stack", from = "src"},
    {include = "modules", from = "src"}
]


[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "0.121.1"
uvicorn = { extras = ["standard"], version = "0.38.0" }
typer = "0.20.0"
rich = "14.2.0"
docker = "^7.1.0"
python-dotenv = "^1.0.1"
pydantic = "^2.7.2"
httpx = "0.28.1"
python-json-logger = "^2.0.7"
azure-devops = "^7.1.0b4"
qdrant-client = "^1.7.3"
litellm = "^1.35.10"
numpy = "^1.26.4"
pyyaml = "^6.0.1"
mcp = "1.2.0"
opentelemetry-sdk = "^1.24.0"
opentelemetry-exporter-otlp = "^1.24.0"
openinference-instrumentation-litellm = "^0.1.0"
opentelemetry-instrumentation-fastapi = "^0.60b1"
trafilatura = "^2.0.0"
sentence-transformers = "^5.2.0"
sqlalchemy = "^2.0.45"
asyncpg = "^0.31.0"
alembic = "^1.17.2"
beautifulsoup4 = "^4.14.3"
langchain-text-splitters = "^1.1.0"
pathspec = "^0.12.1"
aiogram = "^3.10.0"
torch = { version = "2.2.2", source = "pytorch-cpu" }
cryptography = "^46.0.3"
slowapi = "^0.1.9"
pyjwt = "^2.10.1"
python-multipart = "^0.0.22"

[tool.poetry.group.dev.dependencies]
pytest = "9.0.0"
pytest-asyncio = "1.3.0"
ruff = "0.14.4"
black = "25.11.0"
mypy = "^1.10.0"
respx = "^0.22.0"
types-PyYAML = "^6.0.12"
coverage = "7.11.3"
pyyaml = "^6.0.3"
colorama = "^0.4.6"
aiosqlite = "^0.22.1"

[tool.poetry.scripts]
stack = "stack.cli:app"
agent-app = "interfaces.http.app:run"

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = [
    "flows",
    "openwebui",
    ".venv",
]
src = ["src"]

[tool.ruff.lint.isort]
known-first-party = ["core", "modules", "orchestrator", "interfaces", "stack"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "S", "N"]
ignore = ["S101", "S104", "B008", "PLR0912", "PLR0915"]

[tool.ruff.lint.mccabe]
max-complexity = 18

[tool.black]
line-length = 100
target-version = ["py311"]
exclude = '''\
/( 
    \.git
  | \.venv
  | services/agent/\.venv
  | \.venv_temp
  | _build
  | dist
  | tests/.venv
  | src/core/core
  | data
  | services/openwebui/data
)/ 
'''

[tool.mypy]
python_version = "3.11"
mypy_path = ["src"]
# Strictness Flags
check_untyped_defs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
# Robustness
ignore_missing_imports = true 
files = ["src/core", "src/orchestrator", "src/interfaces", "src/stack", "src/modules"]
explicit_package_bases = true

[[tool.mypy.overrides]]
module = ["typer", "typer.*", "yaml", "aiogram", "aiogram.*"]
ignore_errors = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra --import-mode=importlib"
asyncio_mode = "auto"
testpaths = ["../../tests", "src/core/tests", "src/stack/tests"]

[tool.coverage.run]
branch = true
source = ["src"]

[tool.coverage.report]
show_missing = true
skip_empty = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == \"__main__\":",
    "if TYPE_CHECKING:",
]

[tool.coverage.html]
directory = "htmlcov"



[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint.per-file-ignores]
"src/core/models/gemini_cli.py" = ["S603"]
"**/stack/checks.py" = ["S603"]
"**/tests/**/*.py" = ["S105", "S106"]


